@@ -2,21 +2,22 @@ class UsersController < ApplicationController
   respond_to :html, :json, :csv
   respond_to :js, :only => [:update]
   # Privileged controller
-  before_filter :authenticate_admin
+  before_action :authenticate_admin
   # Edit the submitted admin collections based on existing collections in @user's db entry
-  before_filter :update_admin_collections, :only => [:update]
-  
+  before_action :update_admin_collections, :only => [:update]
+
   # GET /users
   def index
-    @users = User.search(params[:q]).order(sort_column + " " + sort_direction).page(params[:page]).per(30)
-    
+    @users = User.order(sort_column + " " + sort_direction).page(params[:page]).per(30)
+    @users = @users.with_query(params[:q]) unless params[:q].blank?
+
     respond_with(@users)
   end
 
   # GET /users/1
   def show
     @user = User.find(params[:id])
-    
+
     respond_with(@user)
   end
 
@@ -27,13 +28,14 @@ class UsersController < ApplicationController
       :umbra_admin_collections => @user.user_collections,
       :umbra_admin => !@user.user_collections.empty?
     }
+
     @user.user_attributes = user_attributes
-    
-    flash[:notice] = t('users.update_success') if @user.save 
+
+    flash[:notice] = t('users.update_success') if @user.save
 
     respond_with(@user)
   end
-  
+
   # DELETE /users/1
   def destroy
     @user = User.find(params[:id])
@@ -41,18 +43,18 @@ class UsersController < ApplicationController
 
     respond_with(@user)
   end
-  
+
   # DELETE /users/clear_patron_data
   def clear_patron_data
     @users = User.all
-    
-    flash[:success] = t('users.clear_patron_data_success') if User.destroy_all("user_attributes not like '%:umbra_admin: true%'")
-        
+
+    flash[:success] = t('users.clear_patron_data_success') if User.non_admin.destroy_all
+
     respond_with(@users, :location => users_url) do |format|
       format.html { redirect_to users_url }
     end
   end
-  
+
   # Implement sort column for User class
   def sort_column
     super "User", "lastname"
