@@ -1,6 +1,6 @@
 //= require jquery
 //= require jquery_ujs
-//= require jquery.ui.all
+//= require jquery-ui
 // Required by Blacklight
 //= require blacklight/blacklight
 //= require nyulibraries
