@@ -1,16 +1,9 @@
 # -*- encoding : utf-8 -*-
-class SolrDocument 
+class SolrDocument
 
   include Blacklight::Solr::Document
 
   self.unique_key = 'resource_id_ss'
-  
-  # The following shows how to setup this blacklight document to display marc documents
-  #extension_parameters[:marc_source_field] = :marc_display
-  #extension_parameters[:marc_format_type] = :marcxml
-  #use_extension( Blacklight::Solr::Document::Marc) do |document|
-  #  document.key?( :marc_display  )
-  #end
 
   # Email uses the semantic field mappings below to generate the body of an email.
   SolrDocument.use_extension( Blacklight::Solr::Document::Email )
@@ -23,9 +16,6 @@ class SolrDocument
   # single valued. See Blacklight::Solr::Document::ExtendableClassMethods#field_semantics
   # and Blacklight::Solr::Document#to_semantic_values
   # Recommendation: Use field names from Dublin Core
-  SolrDocument.use_extension( Blacklight::Solr::Document::DublinCore)    
-  field_semantics.merge!(    
-                        :title => "title_texts",
-                        :description => "description_texts"
-                        )
+  use_extension( Blacklight::Solr::Document::DublinCore)
+
 end
