@@ -1,3 +1,4 @@
-require 'nyulibraries/deploy/capistrano'
+require 'formaggio/capistrano'
 
 set :app_title, "umbra"
+set :rvm_ruby_string, "2.1.3"
