@@ -1,2 +1,2 @@
 set :rails_env, "staging"
-set :branch, "master"
+set(:branch, ENV["GIT_BRANCH"].gsub(/remotes\//,"").gsub(/origin\//,""))
