@@ -1,6 +1,5 @@
-Umbra::Application.routes.draw do
-
-  Blacklight.add_routes(self)
+Rails.application.routes.draw do
+  blacklight_for :catalog
 
   root :to => "catalog#index", :collection => "vbl"
 
@@ -8,7 +7,7 @@ Umbra::Application.routes.draw do
   YAML.load_file( File.join(Rails.root, "config", "repositories.yml") )["Catalog"]["repositories"].each do |coll|
      get "#{coll[0]}" => "catalog#index", :collection => "#{coll[0]}"
   end
-      
+
   scope "admin" do
     resources :records do
       post 'upload', :on => :collection
@@ -16,10 +15,9 @@ Umbra::Application.routes.draw do
     resources :users
     delete "clear_patron_data", :to => "users#clear_patron_data"
   end
-  
-  match 'login', :to => 'user_sessions#new', :as => :login
-  match 'logout', :to => 'user_sessions#destroy', :as => :logout
-  match 'validate', :to => 'user_sessions#validate', :as => :validate
+
+  get 'login', :to => 'user_sessions#new', :as => :login
+  get 'logout', :to => 'user_sessions#destroy', :as => :logout
+  get 'validate', :to => 'user_sessions#validate', :as => :validate
   resources :user_sessions
-  
 end
